# docs/api/openapi.json
{
  "components" : {
    "schemas" : {
...
      "ProblemDetail" : {
        "properties" : {
          "detail" : {
            "type" : "string"
          },
          "instance" : {
            "format" : "uri",
            "type" : "string"
          },
          "properties" : {
            "additionalProperties" : { },
            "type" : "object"
          },
          "status" : {
            "format" : "int32",
            "type" : "integer"
          },
          "title" : {
            "type" : "string"
          },
          "type" : {
            "format" : "uri",
            "type" : "string"
          }
        },
        "type" : "object"
      }
    }
  },
  "info" : {
    "description" : "Create, read and retire short links; visitors are redirected with 302.",
    "title" : "urlshort",
    "version" : "1"
  },
  "openapi" : "3.1.0",
  "paths" : {
...
    "/{code}" : {
      "get" : {
        "operationId" : "redirect",
        "parameters" : [ {
          "in" : "path",
          "name" : "code",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "302" : {
            "description" : "Redirect to the stored url, not cacheable",
            "headers" : {
              "Cache-Control" : {
                "description" : "no-store",
                "schema" : {
                  "type" : "string"
                },
                "style" : "simple"
              },
              "Location" : {
                "description" : "The stored url, byte for byte",
                "schema" : {
                  "type" : "string"
                },
                "style" : "simple"
              }
            }
          },
          "404" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description" : "No link has this code"
          },
          "410" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description" : "The link was retired"
          }
        },
        "summary" : "Open a short link",
        "tags" : [ "redirect-controller" ]
      }
    }
  },
  "servers" : [ {
    "url" : "/"
  } ]
}